Migrate to onemap

+++ server/model/routeInfo.go
@@ -1,7 +1,7 @@
 package model
 
 type RouteInfo struct {
-	Distance int32  `json:"distance"`
-	Duration int64  `json:"duration"`
+	Distance float64  `json:"distance"`
+	Duration float64  `json:"duration"`
 	Polyline string `json:"polyline"`
 }
