+++ server/utils/routeConverter.go
@@ -1,9 +1,9 @@
 package utils
 
-func ConvertMeterToKm(meter int32) float64 {
-	return float64(meter) / 1000.0
+func ConvertMeterToKm(meter float64) float64 {
+	return meter / 1000
 }
 
-func ConvertSecondsToMinutes(seconds int64) float64 {
-	return float64(seconds) / 60.0
+func ConvertSecondsToMinutes(seconds float64) float64 {
+	return seconds / 60
 }
